WIP: Address PR review comments for PrescribedTracer
